# apps/api/internal/httpapi/gen.go
// Package httpapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package httpapi

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Error defines model for Error.
type Error struct {
	// Code Machine-readable code, e.g. invalid_request.
	Code    string `json:"code"`
	Message string `json:"message"`
}

// HistoryResponse defines model for HistoryResponse.
type HistoryResponse struct {
	// Points Price samples, oldest first, ending at the current price.
	Points []float64 `json:"points"`
	Symbol string    `json:"symbol"`
}

// Listing defines model for Listing.
type Listing struct {
	Exchange string `json:"exchange"`
	Name     string `json:"name"`
	Sector   string `json:"sector"`
	Symbol   string `json:"symbol"`
}

// Quote defines model for Quote.
type Quote struct {
	// AsOf When this quote was fetched from the source (UTC).
	AsOf time.Time `json:"asOf"`
	Beta *float64  `json:"beta"`

	// ChangeAbs Day change in dollars, signed.
	ChangeAbs float64 `json:"changeAbs"`

	// ChangePct Day change in percent, signed.
	ChangePct float64 `json:"changePct"`

	// DivYield Percent, e.g. 0.44 means 0.44%.
	DivYield *float64 `json:"divYield"`
	Eps      *float64 `json:"eps"`
	Exchange string   `json:"exchange"`
	High     float64  `json:"high"`
	Low      float64  `json:"low"`

	// MarketCap Raw dollars, not abbreviated.
	MarketCap float64  `json:"marketCap"`
	Name      string   `json:"name"`
	Open      float64  `json:"open"`
	PeRatio   *float64 `json:"peRatio"`
	PrevClose float64  `json:"prevClose"`
	Price     float64  `json:"price"`
	Sector    string   `json:"sector"`
	Symbol    string   `json:"symbol"`

	// Volume Raw share count, not abbreviated.
	Volume     float64 `json:"volume"`
	Week52High float64 `json:"week52High"`
	Week52Low  float64 `json:"week52Low"`
}

// QuotesResponse defines model for QuotesResponse.
type QuotesResponse struct {
	Quotes []Quote `json:"quotes"`
}

// SymbolsResponse defines model for SymbolsResponse.
type SymbolsResponse struct {
	Symbols []Listing `json:"symbols"`
}

// GetQuotesParams defines parameters for GetQuotes.
type GetQuotesParams struct {
	// Symbols Comma-separated ticker symbols, e.g. AAPL,MSFT. Max 50.
	Symbols string `form:"symbols" json:"symbols"`
}

// GetQuoteHistoryParams defines parameters for GetQuoteHistory.
type GetQuoteHistoryParams struct {
	// Points Number of samples. Defaults to 70.
	Points *int `form:"points,omitempty" json:"points,omitempty"`
}

// GetSymbolsParams defines parameters for GetSymbols.
type GetSymbolsParams struct {
	// Query Case-insensitive substring match on symbol or name. Absent or empty returns the full catalog.
	Query *string `form:"query,omitempty" json:"query,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Current quotes for a set of symbols
	// (GET /v1/quotes)
	GetQuotes(c *gin.Context, params GetQuotesParams)
	// Price series for one symbol
	// (GET /v1/quotes/{symbol}/history)
	GetQuoteHistory(c *gin.Context, symbol string, params GetQuoteHistoryParams)
	// Search the symbol directory
	// (GET /v1/symbols)
	GetSymbols(c *gin.Context, params GetSymbolsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetQuotes operation middleware
func (siw *ServerInterfaceWrapper) GetQuotes(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetQuotesParams

	// ------------- Required query parameter "symbols" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "symbols", c.Request.URL.Query(), &params.Symbols, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter symbols: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetQuotes(c, params)
}

// GetQuoteHistory operation middleware
func (siw *ServerInterfaceWrapper) GetQuoteHistory(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "symbol" -------------
	var symbol string

	err = runtime.BindStyledParameterWithOptions("simple", "symbol", c.Param("symbol"), &symbol, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter symbol: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetQuoteHistoryParams

	// ------------- Optional query parameter "points" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "points", c.Request.URL.Query(), &params.Points, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter points: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetQuoteHistory(c, symbol, params)
}

// GetSymbols operation middleware
func (siw *ServerInterfaceWrapper) GetSymbols(c *gin.Context) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params GetSymbolsParams

	// ------------- Optional query parameter "query" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "query", c.Request.URL.Query(), &params.Query, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter query: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetSymbols(c, params)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/v1/quotes", wrapper.GetQuotes)
	router.GET(options.BaseURL+"/v1/quotes/:symbol/history", wrapper.GetQuoteHistory)
	router.GET(options.BaseURL+"/v1/symbols", wrapper.GetSymbols)
}

type GetQuotesRequestObject struct {
	Params GetQuotesParams
}

type GetQuotesResponseObject interface {
	VisitGetQuotesResponse(w http.ResponseWriter) error
}

type GetQuotes200JSONResponse QuotesResponse

func (response GetQuotes200JSONResponse) VisitGetQuotesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuotes400JSONResponse Error

func (response GetQuotes400JSONResponse) VisitGetQuotesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuotes500JSONResponse Error

func (response GetQuotes500JSONResponse) VisitGetQuotesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuotes503JSONResponse Error

func (response GetQuotes503JSONResponse) VisitGetQuotesResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuoteHistoryRequestObject struct {
	Symbol string `json:"symbol"`
	Params GetQuoteHistoryParams
}

type GetQuoteHistoryResponseObject interface {
	VisitGetQuoteHistoryResponse(w http.ResponseWriter) error
}

type GetQuoteHistory200JSONResponse HistoryResponse

func (response GetQuoteHistory200JSONResponse) VisitGetQuoteHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuoteHistory400JSONResponse Error

func (response GetQuoteHistory400JSONResponse) VisitGetQuoteHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuoteHistory404JSONResponse Error

func (response GetQuoteHistory404JSONResponse) VisitGetQuoteHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuoteHistory500JSONResponse Error

func (response GetQuoteHistory500JSONResponse) VisitGetQuoteHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetQuoteHistory503JSONResponse Error

func (response GetQuoteHistory503JSONResponse) VisitGetQuoteHistoryResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetSymbolsRequestObject struct {
	Params GetSymbolsParams
}

type GetSymbolsResponseObject interface {
	VisitGetSymbolsResponse(w http.ResponseWriter) error
}

type GetSymbols200JSONResponse SymbolsResponse

func (response GetSymbols200JSONResponse) VisitGetSymbolsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetSymbols400JSONResponse Error

func (response GetSymbols400JSONResponse) VisitGetSymbolsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetSymbols500JSONResponse Error

func (response GetSymbols500JSONResponse) VisitGetSymbolsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Current quotes for a set of symbols
	// (GET /v1/quotes)
	GetQuotes(ctx context.Context, request GetQuotesRequestObject) (GetQuotesResponseObject, error)
	// Price series for one symbol
	// (GET /v1/quotes/{symbol}/history)
	GetQuoteHistory(ctx context.Context, request GetQuoteHistoryRequestObject) (GetQuoteHistoryResponseObject, error)
	// Search the symbol directory
	// (GET /v1/symbols)
	GetSymbols(ctx context.Context, request GetSymbolsRequestObject) (GetSymbolsResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetQuotes operation middleware
func (sh *strictHandler) GetQuotes(ctx *gin.Context, params GetQuotesParams) {
	var request GetQuotesRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetQuotes(ctx, request.(GetQuotesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetQuotes")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetQuotesResponseObject); ok {
		if err := validResponse.VisitGetQuotesResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetQuoteHistory operation middleware
func (sh *strictHandler) GetQuoteHistory(ctx *gin.Context, symbol string, params GetQuoteHistoryParams) {
	var request GetQuoteHistoryRequestObject

	request.Symbol = symbol
	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetQuoteHistory(ctx, request.(GetQuoteHistoryRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetQuoteHistory")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetQuoteHistoryResponseObject); ok {
		if err := validResponse.VisitGetQuoteHistoryResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetSymbols operation middleware
func (sh *strictHandler) GetSymbols(ctx *gin.Context, params GetSymbolsParams) {
	var request GetSymbolsRequestObject

	request.Params = params

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetSymbols(ctx, request.(GetSymbolsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetSymbols")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetSymbolsResponseObject); ok {
		if err := validResponse.VisitGetSymbolsResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}